requirements.txt
streamlit
pandas
plotly
supabase
requests
Pillow

# Optional (for future DB support)
# psycopg2-binary
